refactor: ExperienceCard 컴포넌트 분리
